fixing dragging bug and moved dragging state viewmodel
